fix: splines drawing at (0,0) if drawn before 5 seconds
