+++ src/toolboxes/background/Canvas.tsx
@@ -39,13 +39,13 @@ export class BackgroundCanvas extends Component<Props> {
     this.image.crossOrigin = "anonymous";
 
     // Draw the image once loaded
-    this.image.onload = this.redrawImage;
+    this.image.onload = () => {this.redrawImage; this.props.updateImageDimensions(this.image.width, this.image.height);}
     this.image.src = this.props.imgSrc;
 
-    //TODO: remove
-    setTimeout(() => {
-      this.props.updateImageDimensions(this.image.width, this.image.height);
-    }, 5000);
+    // //TODO: remove
+    // setTimeout(() => {
+    //   this.props.updateImageDimensions(this.image.width, this.image.height);
+    // }, 5000);
   };
 
   componentDidMount = (): void => {
