+++ src/toolboxes/spline/Canvas.tsx
@@ -3,7 +3,8 @@ import { Component } from "react";
 
 // import { SplineVector } from "./interfaces";
 import { BaseCanvas, Props as BaseProps } from "../../baseCanvas";
-import { Annotations, canvasToImage, imageToCanvas, imageToOriginalCanvas } from "../../annotation";
+import { Annotations } from "../../annotation";
+import { canvasToImage, imageToCanvas, imageToOriginalCanvas } from "../../baseCanvas";
 import { XYPoint } from "../../annotation/interfaces";
 
 interface Props extends BaseProps {
@@ -23,8 +24,8 @@ export class SplineCanvas extends Component<Props> {
     super(props);
   }
 
-  drawSplineVector = (currentSplineVector: XYPoint[]) => {
-    if (currentSplineVector.length < 2) return;
+  drawSplineVector = (splineVector: XYPoint[]) => {
+    if (splineVector.length < 2) return;
 
     const { canvasContext: context, canvas } = this.baseCanvas;
     const lineWidth = 1.5;
@@ -37,14 +38,14 @@ export class SplineCanvas extends Component<Props> {
     context.beginPath();
 
     // Go to the first point
-    let firstPoint:XYPoint = currentSplineVector[0];
+    let firstPoint: XYPoint = splineVector[0];
     firstPoint = imageToOriginalCanvas(firstPoint.x, firstPoint.y, this.props.imageWidth, this.props.imageHeight, this.props.scaleAndPan);
 
     context.moveTo(firstPoint.x, firstPoint.y);
 
     // Draw each point by taking our raw coordinates and applying the transform so they fit on our canvas
     let nextPoint;
-    for (const { x, y } of currentSplineVector) {
+    for (const { x, y } of splineVector) {
       nextPoint = imageToOriginalCanvas(x, y, this.props.imageWidth, this.props.imageHeight, this.props.scaleAndPan);
       context.lineTo(nextPoint.x, nextPoint.y);
     }
